Fix TS build and add index.ts
